refactor: moved faceit_util.ts to widget/src/utils
